Add DJ claim feature and restrict login
DJs can now only log into DJs that they have claimed. When registering a
new DJ through the DJ interface, a claim is automatically created for
the current user.

The workflow to create and process claims for DJs that already exist or
are created through the admin interface has not yet been built.