Remove Redis caching and instead use HTTP caching
The existing Redis caching was unnecessarily complicated and
reimplemented something that reverse proxies can already do for us
better. This change adds Cache-Control and Last-Modified headers where
appropriate so that caching can be handled by a reverse proxy.

As with the built-in caching, the playlists cache will be invalidated as
needed to ensure we don't serve stale data. As a result, cached
playlists requests will still hit the app server, but the conditional
request logic will prevent those requests from hitting the database.