Don't return 400 errors for search/autocomplete
There are a few cases where Trackman clients may send searches that are
empty. Instead of returning a 400 error, just return a 200 error with no
results.

For autocomplete, we also do this if a field that we aren't able to
search on (e.g. vinyl) is provided. This is actually in line with the
existing search be behavior, which ignores fields it cannot search on.
In the future though, we may change this behavior.

This was caught by our Sentry error logging.