Only update extend autologout status when on air
The Trackman client makes an API call to get the current status of the
"extend autologout" checkbox during initialization. If a DJ is not
currently on air (and thus the djsetId is null), this would result in a
403. Instead, we should only make this API call if djsetId is not null
when initializing Trackman, meaning that the DJ is already on air and we
have a DJSet available to use.