Move Redis access to a KVStore object
This will more easily allow us to replace Redis if necessary. The
KVStore class also provides functionality to cast values to particular
types, which is helpful since Redis always returns bytes.

In the future, it might make sense to store the type alongside the value
to both eliminate the need to specify a type and ensure that we don't
cast to the wrong type.