Add logged out page
Since all non-local requests require login, we need a page to display to
the user after they log out successfully. Otherwise, they will be
redirected back to a login-protected resource, which may result in them
being automatically logged in again to the application.

This separate page that confirms that their logout was successful should
help avoid some confusion by avoiding this workflow and instead
displaying a blank page with a flashed message that confirms that their
logout was successful.