Library: add "MusicBrainz Helper" for artists
This feature allows for a MusicBrainz artist MBID to be entered (or
selected) on the artist page. From there, the user is presented with a
list of unique artist/album combinations and a dropdown to select a
corresponding release group.

If a release group is selected, the album and release group MBIDs are
updated for all tracks matching the search.

This functionality is not perfect as it does not handle cases where the
album names are identical across multiple albums, but it should improve
the workflow for doing this quite a bit.

Fixes #43.