Add new /api/tracklog/add call
This API call automatically determines whether a track already exists or
if one needs to be created. It eliminates the need for the Trackman UI
to keep track of track IDs. (It still can, but creating a track ahead of
time is no longer required.) This will make it easier to implement
alternative options for search.