Try to match tracks by recording MBID first
This updates the find_or_add_track function to query the database for
tracks by recording MBID first, if that field is present. If the recording
MBID field is present but does not match an existing track in the
database, a new track will always be created.

Any deduplication with existing tracks with good metadata but without a
recording MBID is out of scope for this particular function at this
time.